Do not cut 32bit value down to 16bit
handledata contains a 16bit char handle in the lower bits and
may contains a 16bit descriptor handle in its upper 16 bits.
When doing blob read requests for descriptors the upper 16 bit
may be cut off and the descriptor read turns into a characteristic
read. In turn this switches the internal
state of the service discovery state machine causing an endless loop
between descriptor and characteristic reads.

Task-number: QTBUG-47028

+++ src/bluetooth/qlowenergycontroller_bluez.cpp
@@ -1222,7 +1222,7 @@ void QLowEnergyControllerPrivate::readServiceValues(
     starting the next read request.
  */
 void QLowEnergyControllerPrivate::readServiceValuesByOffset(
-        quint16 handleData, quint16 offset, bool isLastValue)
+        uint handleData, quint16 offset, bool isLastValue)
 {
     const QLowEnergyHandle charHandle = (handleData & 0xffff);
     const QLowEnergyHandle descriptorHandle = ((handleData >> 16) & 0xffff);
