+++ src/bluetooth/qlowenergycontroller_p.h
@@ -173,7 +173,7 @@ private:
     void sendReadValueRequest(QLowEnergyHandle attributeHandle, bool isDescriptor);
     void readServiceValues(const QBluetoothUuid &service,
                            bool readCharacteristics);
-    void readServiceValuesByOffset(quint16 handleData, quint16 offset,
+    void readServiceValuesByOffset(uint handleData, quint16 offset,
                                    bool isLastValue);
 
     void discoverServiceDescriptors(const QBluetoothUuid &serviceUuid);
